generate migration command
